+++ src/app/page.tsx
@@ -1,12 +1,12 @@
 import Service from "@/components/layout/Service";
-import NicknameManager from "./(auth)/login/components/NicknameManager";
+// import NicknameManager from "./(auth)/login/components/NicknameManager";
 import { Modal } from "@/components/shared/Modal";
 
 export default function Home() {
   return (
     <>
       <Service />
-      <NicknameManager />
+      {/* <NicknameManager /> */}
       <Modal />
     </>
   );
